pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["v3code <vitalik.w555@gmail.com>"]
readme = "README.md"
packages = [{include = "page_analyzer"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
gunicorn = "^20.1.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.6"


[tool.poetry.group.dev.dependencies]
pylama = {extras = ["all"], version = "^8.4.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
